+++ webui/react/src/pages/Authentication.tsx
@@ -1,14 +1,15 @@
 import { Button, Form, Input } from 'antd';
 import { message } from 'antd';
+import axios from 'axios';
 import queryString from 'query-string';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { useHistory, useLocation } from 'react-router';
 import { Redirect } from 'react-router-dom';
 
 import Icon from 'components/Icon';
 import Logo, { LogoTypes } from 'components/Logo';
 import Spinner from 'components/Spinner';
-import Auth from 'contexts/Auth';
+import Auth, { updateAuth } from 'contexts/Auth';
 import handleError, { ErrorType } from 'ErrorHandler';
 import { login, logout } from 'services/api';
 import { Credentials } from 'types';
@@ -28,10 +29,19 @@ const Authentication: React.FC<WithSearch<{}>> = (props: WithSearch<{}>) => {
   const location = useLocation();
   const auth = Auth.useStateContext();
   const setAuth = Auth.useActionContext();
+  const [ isLoading, setIsLoading ] = useState(true);
 
   const queries: Queries = queryString.parse(props.location.search);
 
   const redirectUri= queries.redirect || DEFAULT_REDIRECT;
+
+  useEffect(() => {
+    const source = axios.CancelToken.source();
+    updateAuth(setAuth, source.token).then(() => setIsLoading(false));
+    return (): void => {
+      source.cancel();
+    };
+  }, [ setAuth ]);
 
   const isLogout = location.pathname.endsWith('logout');
   if (isLogout) {
@@ -39,7 +49,6 @@ const Authentication: React.FC<WithSearch<{}>> = (props: WithSearch<{}>) => {
     logout({});
     setAuth({ type: Auth.ActionType.Reset });
     history.push('/det/login' + props.location.search);
-    return <Spinner fullPage />;
   } else {
     console.log('is in login page');
   }
@@ -96,9 +105,11 @@ const Authentication: React.FC<WithSearch<{}>> = (props: WithSearch<{}>) => {
 
   );
 
+  if (isLogout || isLoading) return <Spinner fullPage />;
+  if (auth.isAuthenticated) return <Redirect to={redirectUri} />;
+
   return (
     <div className={css.base}>
-      {auth.isAuthenticated && <Redirect to={redirectUri} />}
       <div className={css.content}>
         {/* DISCUSSION what if we didn't need to add the logo classname and was able to
         target logo on its own using component name easily in module.scss */}
